Fixes aggrid header dependency cycle
