+++ src/components/custom-aggrid/custom-aggrid-filters/hooks/use-custom-aggrid-comparator-filter.ts
@@ -4,7 +4,6 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-import { FILTER_DATA_TYPES } from '../../custom-aggrid-header.type';
 import { ChangeEvent, useMemo } from 'react';
 import { useSnackMessage } from '@gridsuite/commons-ui';
 import { SelectChangeEvent } from '@mui/material/Select/SelectInput';
@@ -13,6 +12,7 @@ import { useCustomAggridFilter } from './use-custom-aggrid-filter';
 import { GridApi } from 'ag-grid-community';
 import { computeTolerance } from '../utils/filter-tolerance-utils';
 import { FilterParams } from '../../../../types/custom-aggrid-types';
+import { FILTER_DATA_TYPES } from '../custom-aggrid-filter.type';
 
 export const useCustomAggridComparatorFilter = (api: GridApi, colId: string, filterParams: FilterParams) => {
     const { dataType = FILTER_DATA_TYPES.TEXT } = filterParams;
