+++ src/components/custom-aggrid/custom-aggrid-filters/utils/filter-tolerance-utils.ts
@@ -6,8 +6,8 @@
  */
 import { isNumber } from 'mathjs';
 import { countDecimalPlaces, countDecimalPlacesFromString } from '../../../../utils/rounding';
-import { FILTER_NUMBER_COMPARATORS, UNDISPLAYED_FILTER_NUMBER_COMPARATORS } from '../../custom-aggrid-header.type';
 import { FilterConfig } from '../../../../types/custom-aggrid-types';
+import { FILTER_NUMBER_COMPARATORS, UNDISPLAYED_FILTER_NUMBER_COMPARATORS } from '../custom-aggrid-filter.type';
 
 /**
  * Compute the tolerance that should be applied when comparing filter values to database values
