+++ src/components/results/loadflow/load-flow-result-tab.tsx
@@ -29,7 +29,6 @@ import {
     mappingTabs,
     useFetchFiltersEnums,
 } from './load-flow-result-utils';
-import { FILTER_DATA_TYPES, FILTER_TEXT_COMPARATORS } from 'components/custom-aggrid/custom-aggrid-header.type';
 import { LimitViolationResult } from './limit-violation-result';
 import { NumberCellRenderer, StatusCellRender } from '../common/result-cell-renderers';
 import ResultsGlobalFilter, { Filter } from '../common/results-global-filter';
@@ -43,6 +42,10 @@ import { mapFieldsToColumnsFilter } from '../../../utils/aggrid-headers-utils';
 import { loadflowResultInvalidations } from '../../computing-status/use-all-computing-status';
 import { FilterType } from '../common/utils';
 import { useNodeData } from 'components/use-node-data';
+import {
+    FILTER_DATA_TYPES,
+    FILTER_TEXT_COMPARATORS,
+} from '../../custom-aggrid/custom-aggrid-filters/custom-aggrid-filter.type';
 
 const styles = {
     flexWrapper: {
