+++ src/components/spreadsheet/custom-columns/use-custom-column.ts
@@ -8,7 +8,7 @@ import { useCallback, useMemo } from 'react';
 import { AppState } from 'redux/reducer';
 import { useSelector } from 'react-redux';
 import { CustomColumnMenu } from '../../custom-aggrid/custom-column-menu';
-import { COLUMN_TYPES, CustomColDef } from '../../custom-aggrid/custom-aggrid-header.type';
+import { COLUMN_TYPES } from '../../custom-aggrid/custom-aggrid-header.type';
 import { limitedEvaluate } from './math';
 import { ColDef, ValueGetterParams } from 'ag-grid-community';
 import {
@@ -19,6 +19,7 @@ import {
 } from '../config/common-column-definitions';
 import { validateFormulaResult } from './formula-validator';
 import { ColumnDefinition } from '../config/spreadsheet.type';
+import { CustomColDef } from '../../custom-aggrid/custom-aggrid-filters/custom-aggrid-filter.type';
 
 export function useCustomColumn(tabIndex: number) {
     const tableDefinition = useSelector((state: AppState) => state.tables.definitions[tabIndex]);
