+++ src/components/spreadsheet/table-wrapper.tsx
@@ -24,7 +24,6 @@ import CustomColumnsConfig from './custom-columns/custom-columns-config';
 import { AppState, CurrentTreeNode } from '../../redux/reducer';
 import { AgGridReact } from 'ag-grid-react';
 import { ColumnMovedEvent, ColumnState, RowClickedEvent } from 'ag-grid-community';
-import { CustomColDef } from '../custom-aggrid/custom-aggrid-header.type';
 import { SpreadsheetEquipmentType } from './config/spreadsheet.type';
 import SpreadsheetSave from './spreadsheet-save';
 import CustomColumnsNodesConfig from './custom-columns/custom-columns-nodes-config';
@@ -36,6 +35,7 @@ import { useEquipmentModification } from './equipment-modification/use-equipment
 import { useSpreadsheetGsFilter } from './use-spreadsheet-gs-filter';
 import { updateTableDefinition } from 'redux/actions';
 import { NodeType } from '../graph/tree-node.type';
+import { CustomColDef } from '../custom-aggrid/custom-aggrid-filters/custom-aggrid-filter.type';
 
 const styles = {
     table: (theme: Theme) => ({
